requirements.txt.txt
# Revoda Backend Dependencies
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
asyncpg==0.29.0         # PostgreSQL async driver
psycopg2-binary==2.9.9  # For migrations

# Cache
redis[asyncio]==5.0.4

# Auth & Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4             # Password / token hashing

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1          # .env config

# HTTP
httpx==0.27.0                     # Async HTTP client (for partner webhooks)

# Email / alerts
sendgrid==6.11.0                  # Email escalation alerts

# SMS (Africa's Talking for feature phone users in Nigeria)
africastalking==1.2.7

# File uploads / S3
boto3==1.34.88                    # AWS S3 for media storage

# Background jobs
celery[redis]==5.4.0              # For scheduled hotspot scans
flower==2.0.1                     # Celery monitoring

# Data / Analysis
pandas==2.2.2                     # Data export / reporting
geopy==2.4.1                      # Geocoding utility

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                     # For API test client
